Add simple evaluation function

+++ Src/evaluation.h
@@ -6,10 +6,10 @@ class Evaluation
 {
 public:
     static const std::map<PieceName, int> pieceNameToValue;
-    static int EvaluateMaterial(const Board &board, PieceColors sideToEvaluate);
+    static float EvaluateMaterial(const Board &board, PieceColors sideToEvaluate);
     static int EvaluateHangingPieces(const Board &board, PieceColors sideToEvaluate);
     static int EvaluateDefendedPieces(const Board &board, PieceColors sideToEvaluate);
     static int EvaluateKingSafety(const Board &board, PieceColors sideToEvaluate);
-    static int EvaluateSide(const Board &board, PieceColors sideToEvaluate);
-    static int Evaluate(const Board &board);
+    static float EvaluateSide(const Board &board, PieceColors sideToEvaluate);
+    static float Evaluate(const Board &board);
 };
